bandwidth v1.2.2 Bandwidth.Resources.Account

The Account resource allows you to retrieve your current balance, transaction list, account type and all elements related to your platform account.

Bandwidth Docs

Functions

self(client)

self(Client.t()) :: Client.response()

Get information about your account.

Example:

case Bandwidth.Resources.Account.self(client) do
  {:ok, {200, account, _}} -> IO.inspect account
  {:error, reason}         -> IO.puts "Error: #{reason}"
end

Bandwidth Docs

transactions(client, params \\ [])

transactions(Client.t(), Client.params()) :: Client.response()

Get a list of the transactions made to your account.

Example:

case Bandwidth.Resources.Account.transactions(client, type: "charge") do
  {:ok, {200, transactions, _}} -> IO.inspect transactions
  {:error, reason}              -> IO.puts "Error: #{reason}"
end

Bandwidth Docs